added a fixed size array, simplified the code
