+++ savintsev.ilya/T1/shape.hpp
@@ -20,7 +20,7 @@ namespace savintsev
   void scaleRelativeTo(Shape * rhs, point_t p, double k);
   void scaleAllRelativeTo(Shape ** rhs, size_t amt, point_t p, double k);
   void printSumAreaAndBorders(std::ostream & out, Shape ** rhs, size_t amt);
-  void destroyShapes(Shape ** rhs, size_t amt);
+  int actWithShpByDesc(char * desc, Shape ** rhs, size_t & amt, point_t & p, double & k);
 
 }
 #endif
